Separate CesiumFeaturesMetadataDescription into own file
